RavenDB-21646 : address PR comments
